Split out brushes and drag events into their own directories.
