+++ mapper/brushes/index.js
@@ -1,0 +1,4 @@
+export * from "./brush.js";
+export * from "./add_brush.js";
+export * from "./delete_brush.js";
+export * from "./select_brush.js";
