+++ mapper/mapper.js
@@ -1,586 +1,9 @@
 import { HookContainer } from "./hook_container.js";
-import { Vector3, Box3, Line3 } from "./geometry.js";
+import { Vector3, Box3 } from "./geometry.js";
 import { asyncFrom, mod } from "./utils.js";
-import { BulkAction, DrawPathAction, TranslateAction, RemoveAction } from "./actions/index.js";
-
-class Brush {
-	constructor(context) {
-		this.context = context;
-
-		this.size = 1;
-		this.maxSize = 10;
-		this.lastSizeChange = performance.now();
-	}
-
-	getDescription() {
-		throw "description not implemented";
-	}
-
-	getRadius() {
-		return this.size * 32;
-	}
-
-	increment() {}
-
-	decrement() {}
-
-	shrink() {
-		this.size = Math.max(1, this.size - 1);
-		this.lastSizeChange = performance.now();
-	}
-
-	enlarge() {
-		this.size = Math.min(this.maxSize, this.size + 1);
-		this.lastSizeChange = performance.now();
-	}
-
-	sizeRecentlyChanged() {
-		return performance.now() - this.lastSizeChange < 1000;
-	}
-
-	async drawAsCircle(context, position) {
-		context.beginPath();
-		context.arc(position.x, position.y, this.getRadius(), 0, 2 * Math.PI, false);
-		context.strokeStyle = "white";
-		context.stroke();
-	}
-
-	async draw(context, position) {
-		await this.drawAsCircle(context, position);
-	}
-
-	async trigger(where, mouseDragEvent) {
-		where;
-		mouseDragEvent;
-	}
-
-	async activate(where) {
-		where;
-	}
-}
-
-class DeleteBrush extends Brush {
-	getDescription() {
-		return `Delete (size ${this.size})`;
-	}
-
-	async activate(where) {
-		return new CumulativeDrawEvent(this.context, where);
-	}
-
-	async * getNodesInBrush(brushPosition) {
-		for await (const nodeRef of this.context.drawnNodes()) {
-			if(this.context.mapPointToCanvas((await nodeRef.center())).subtract(brushPosition).length() <= this.getRadius() && await nodeRef.getPNumber("radius") > 0) {
-				yield nodeRef;
-			}
-		}
-	}
-
-	async draw(context, position) {
-		if(this.context.isKeyDown("Control") || this.sizeRecentlyChanged()) {
-			await this.drawAsCircle(context, position);
-		}
-	}
-
-	async triggerAtPosition(brushPosition) {
-		let toRemove;
-
-		if(this.context.isKeyDown("Control")) {
-			toRemove = await asyncFrom(this.getNodesInBrush(brushPosition));
-		}
-		else {
-			let selection;
-
-			if(this.context.isKeyDown("Shift")) {
-				selection = await Selection.fromNodeIds(this.context, this.context.hoverSelection.parentNodeIds);
-			}
-			else {
-				selection = this.context.hoverSelection;
-			}
-
-			toRemove = selection.getOrigins();
-		}
-
-		return new RemoveAction(this.context, {nodeRefs: toRemove});
-	}
-
-	async triggerOnPath(path) {
-		const actions = [];
-		for(const vertex of path.vertices()) {
-			actions.push(await this.triggerAtPosition(vertex));
-		}
-		return new BulkAction(this.context, {actions: actions});
-	}
-
-	async trigger(path) {
-		return await this.context.performAction(await this.triggerOnPath(path));
-	}
-}
-
-class NodeAddBrush extends Brush {
-	constructor(context) {
-		super(context);
-
-		this.nodeTypeIndex = 1;
-		this.nodeTypes = ["water", "grass", "forest", "mountain"];
-	}
-
-	getDescription() {
-		return `Place ${this.getNodeType()} (size ${this.size})`;
-	}
-
-	getNodeType() {
-		return this.nodeTypes[this.nodeTypeIndex];
-	}
-
-	increment() {
-		this.nodeTypeIndex = this.nodeTypeIndex + 1;
-		this.wrapIndex();
-	}
-
-	decrement() {
-		this.nodeTypeIndex = this.nodeTypeIndex - 1;
-		this.wrapIndex();
-	}
-
-	wrapIndex() {
-		const len = this.nodeTypes.length;
-		this.nodeTypeIndex = (len == 0) ? -1 : mod(this.nodeTypeIndex, len);
-	}
-
-	async trigger(path, mouseDragEvent) {
-		const drawPathActionOptions = {
-			path: path,
-			radius: this.getRadius(),
-			nodeType: this.getNodeType(),
-			scrollOffset: this.context.scrollOffset,
-			fullCalculation: mouseDragEvent.done,
-		};
-
-		const selectionParent = await mouseDragEvent.getSelectionParent();
-		if(selectionParent && await selectionParent.getPString("type") === this.getNodeType()) {
-			drawPathActionOptions.parent = selectionParent;
-		}
-
-		return await this.context.performAction(new DrawPathAction(this.context, drawPathActionOptions));
-	}
-
-	async activate(where) {
-		return new DrawEvent(this.context, where);
-	}
-}
-
-class NodeSelectBrush extends Brush {
-	constructor(context) {
-		super(context);
-	}
-
-	getDescription() {
-		return "Select/Move";
-	}
-
-	async draw(context, position) {
-		context;
-		position;
-	}
-
-	async activate(where) {
-		if(!this.context.hoveringOverSelection()) {
-			if(this.context.hoverSelection.exists()) {
-				let newSelection = null;
-
-				if(this.context.isKeyDown("Shift")) {
-					newSelection = await Selection.fromNodeIds(this.context, this.context.hoverSelection.parentNodeIds);
-				} else {
-					newSelection = this.context.hoverSelection;
-				}
-
-				if(newSelection !== null) {
-					if(this.context.isKeyDown("Control")) {
-						this.context.selection = await this.context.selection.joinWith(newSelection);
-					}
-					else {
-						this.context.selection = newSelection;
-					}
-				}
-			}
-			else {
-				this.context.selection = new Selection(this.context, []);
-			}
-		}
-
-		if(this.context.hoveringOverSelection()) {
-			return new DragEvent(this.context, where, this.context.selection.getOrigins());
-		}
-		else {
-			this.context.selection = new Selection(this, []);
-		}
-	}
-}
-
-class Selection {
-	constructor(context, nodeIds) {
-		this.context = context;
-		this.originIds = new Set(nodeIds);
-		this.parentNodeIds = new Set();
-		this.selectedNodeIds = new Set(this.originIds);
-		this.childNodeIds = new Set();
-		this.siblingNodeIds = new Set();
-		this.directSelectedNodeIds = new Set();
-	}
-
-	static async fromNodeIds(context, nodeIds) {
-		const selection = new Selection(context, nodeIds);
-		await selection.update();
-		return selection;
-	}
-
-	static async fromNodeRefs(context, nodeRefs) {
-		return await Selection.fromNodeIds(context, nodeRefs.map((nodeRef) => nodeRef.id));
-	}
-
-	async joinWith(other) {
-		return Selection.fromNodeRefs(this.context, [...this.getOrigins(), ...other.getOrigins()]);
-	}
-
-	async update() {
-		const selectedNodeIds = new Set(this.originIds);
-		const parentNodeIds = new Set();
-		const childNodeIds = new Set();
-		const siblingNodeIds = new Set();
-		const directSelectedNodeIds = new Set(this.originIds);
-
-		for(const nodeRef of this.getOrigins()) {
-			for await (const childNodeRef of nodeRef.getAllDescendants()) {
-				selectedNodeIds.add(childNodeRef.id);
-				directSelectedNodeIds.add(childNodeRef.id);
-				childNodeIds.add(childNodeRef.id);
-			}
-
-			const parent = await nodeRef.getParent();
-			if(parent) {
-				selectedNodeIds.add(parent.id);
-				parentNodeIds.add(parent.id);
-				directSelectedNodeIds.add(parent.id);
-				for await (const siblingNodeRef of parent.getAllDescendants()) {
-					siblingNodeIds.add(siblingNodeRef.id);
-					selectedNodeIds.add(siblingNodeRef.id);
-				}
-			}
-		}
-
-		this.originIds.forEach((id) => siblingNodeIds.delete(id));
-		parentNodeIds.forEach((id) => siblingNodeIds.delete(id));
-		childNodeIds.forEach((id) => siblingNodeIds.delete(id));
-
-		this.parentNodeIds = parentNodeIds;
-		this.selectedNodeIds = selectedNodeIds;
-		this.childNodeIds = childNodeIds;
-		this.siblingNodeIds = siblingNodeIds;
-		this.directSelectedNodeIds = directSelectedNodeIds;
-	}
-
-	hasNodeRef(nodeRef) {
-		return this.selectedNodeIds.has(nodeRef.id);
-	}
-
-	nodeRefIsOrigin(nodeRef) {
-		return this.originIds.has(nodeRef.id);
-	}
-
-	nodeRefIsParent(nodeRef) {
-		return this.parentNodeIds.has(nodeRef.id);
-	}
-
-	nodeRefIsChild(nodeRef) {
-		return this.childNodeIds.has(nodeRef.id);
-	}
-
-	nodeRefIsSibling(nodeRef) {
-		return this.siblingNodeIds.has(nodeRef.id);
-	}
-
-	getOrigins() {
-		return Array.from(this.originIds.values()).map((id) => this.context.mapper.backend.getNodeRef(id));
-	}
-
-	exists() {
-		return this.originIds.size > 0;
-	}
-
-	contains(other) {
-		for(const nodeId of other.directSelectedNodeIds) {
-			if(!this.directSelectedNodeIds.has(nodeId)) {
-				return false;
-			}
-		}
-		return true;
-	}
-}
-
-class Path {
-	constructor(startPoint) {
-		this.lines = [];
-		this.origin = startPoint;
-		this.at = Vector3.ZERO;
-	}
-
-	mapOrigin(f) {
-		const path = new Path(f(this.origin));
-		path.lines = this.lines;
-		path.at = this.at;
-		return path;
-	}
-
-	mapLines(f) {
-		const path = new Path(this.origin);
-		path.lines = this.lines.map((line) => line.map(f));
-		path.at = f(this.at);
-	}
-
-	withBisectedLines(radius) {
-		const path = new Path(this.origin);
-
-		function addBisectedLine(line) {
-			if(line.distance() >= radius) {
-				const middle = line.a.add(line.b).divideScalar(2);
-				const lineA = new Line3(line.a, middle);
-				const lineB = new Line3(middle, line.b);
-				addBisectedLine(lineA);
-				addBisectedLine(lineB);
-			}
-			else {
-				path.lines.push(line);
-			}
-		}
-
-		for(const line of this.lines) {
-			addBisectedLine(line);
-		}
-
-		path.at = this.at;
-		return path;
-	}
-
-	next(nextPoint) {
-		const nextRelativePoint = nextPoint.subtract(this.origin);
-		if(this.at.subtract(nextRelativePoint).lengthSquared() > 0) {
-			this.lines.push(new Line3(this.at, nextRelativePoint));
-			this.at = nextRelativePoint;
-		}
-	}
-
-	lastLine() {
-		const lastLine = this.lines[this.lines.length - 1];
-		return lastLine ? lastLine : Line3.ZERO;
-	}
-
-	lastVertex() {
-		return this.lastLine().b.add(this.origin);
-	}
-
-	pop() {
-		const lastLine = this.lines.pop();
-		return lastLine ? lastLine : Line3.ZERO;
-	}
-
-	push(line) {
-		return this.lines.push(line);
-	}
-
-	* vertices() {
-		yield this.at.add(this.origin);
-		for(const line of this.lines) {
-			yield line.b.add(this.origin);
-		}
-	}
-
-	getCenter() {
-		const vertices = Array.from(this.vertices());
-		let sum = Vector3.ZERO;
-		for(const vertex of vertices) {
-			sum = sum.add(vertex);
-		}
-		return sum.divideScalar(vertices.length);
-	}
-
-	getRadius() {
-		const center = this.getCenter();
-		let furthest = this.getCenter();
-		for(const vertex of this.vertices()) {
-			if(vertex.subtract(center).lengthSquared() >= furthest.subtract(center).lengthSquared()) {
-				furthest = vertex;
-			}
-		}
-		return furthest.subtract(center).length();
-	}
-
-	asMostRecent() {
-		const lastLine = this.lastLine();
-		const path = new Path(this.origin.add(lastLine.a));
-		path.next(this.origin.add(lastLine.b));
-		return path;
-	}
-}
-
-class MouseDragEvent {
-	constructor(context, startPoint) {
-		this.context = context;
-		this.path = new Path(startPoint);
-		this.done = false;
-		this.hoverSelection = this.context.hoverSelection;
-	}
-
-	next(nextPoint) {
-		this.path.next(nextPoint);
-	}
-
-	end(endPoint) {
-		this.path.next(endPoint);
-		this.done = true;
-	}
-
-	cancel() {}
-
-	async getSelectionParent() {
-		if(this.hoverSelection.exists()) {
-			for(const origin of this.hoverSelection.getOrigins()) {
-				const parent = await origin.getParent();
-				if(parent && !(await origin.hasChildren())) {
-					return parent;
-				}
-				else {
-					return origin;
-				}
-			}
-		}
-
-		return null;
-	}
-}
-
-class DrawEvent extends MouseDragEvent {
-	constructor(context, startPoint) {
-		super(context, startPoint);
-
-		this.undoActions = [];
-	}
-
-	getUndoAction() {
-		return new BulkAction(this.context, {
-			actions: this.undoActions.splice(0, this.undoActions.length).reverse(),
-		});
-	}
-
-	async next(nextPoint) {
-		super.next(nextPoint);
-
-		this.undoActions.push(await this.trigger(this.path.asMostRecent()));
-	}
-
-	async end(endPoint) {
-		super.end(endPoint);
-
-		await this.clear();
-
-		this.undoActions.push(await this.trigger(this.path));
-
-		this.context.pushUndo(this.getUndoAction());
-	}
-
-	async clear() {
-		return await this.context.performAction(this.getUndoAction(), false);
-	}
-
-	async trigger(path) {
-		return await this.context.brush.trigger(path, this);
-	}
-
-	cancel() {
-		this.end(this.path.lastVertex());
-	}
-}
-
-class CumulativeDrawEvent extends MouseDragEvent {
-	constructor(context, startPoint) {
-		super(context, startPoint);
-
-		this.undoActions = [];
-	}
-
-	getUndoAction() {
-		return new BulkAction(this.context, {
-			actions: this.undoActions.splice(0, this.undoActions.length).reverse(),
-		});
-	}
-
-	async next(nextPoint) {
-		super.next(nextPoint);
-
-		this.undoActions.push(await this.trigger(this.path.asMostRecent()));
-	}
-
-	async end(endPoint) {
-		super.end(endPoint);
-
-		this.undoActions.push(await this.trigger(this.path.asMostRecent()));
-
-		this.context.pushUndo(this.getUndoAction());
-	}
-
-	async trigger(path) {
-		return await this.context.brush.trigger(path, this);
-	}
-
-	cancel() {
-		this.end(this.path.lastVertex());
-	}
-}
-
-class DragEvent extends MouseDragEvent {
-	constructor(context, startPoint, nodeRefs) {
-		super(context, startPoint);
-
-		this.nodeRefs = nodeRefs;
-
-		this.undoActions = [];
-	}
-
-	getUndoAction() {
-		return new BulkAction(this.context, {
-			actions: this.undoActions.splice(0, this.undoActions.length).reverse(),
-		});
-	}
-
-	async next(nextPoint) {
-		super.next(nextPoint);
-
-		for(const nodeRef of this.nodeRefs) {
-			this.undoActions.push(await this.context.performAction(new TranslateAction(this.context, {
-				nodeRef: nodeRef,
-				offset: this.path.lastLine().vector(),
-			}), false));
-		}
-	}
-
-	async end(endPoint) {
-		this.next(endPoint);
-
-		this.context.pushUndo(this.getUndoAction());
-	}
-
-	cancel() {
-		this.context.performAction(this.getUndoAction(), false);
-	}
-}
-
-class PanEvent extends MouseDragEvent {
-	next(nextPoint) {
-		super.next(nextPoint);
-		this.context.scrollOffset = this.context.scrollOffset.subtract(this.path.lastLine().vector());
-		this.context.recalculateTilesViewport();
-	}
-}
+import { DeleteBrush, AddBrush, SelectBrush } from "./brushes/index.js";
+import { PanEvent } from "./drag_events/index.js";
+import { Selection } from "./selection.js";
 
 /** A render context of a mapper into a specific element.
  * Handles keeping the UI connected to an element on a page.
@@ -625,7 +48,7 @@ class RenderContext {
 
 		this.scrollOffset = Vector3.ZERO;
 
-		this.brush = new NodeAddBrush(this);
+		this.brush = new AddBrush(this);
 
 		this.hoverSelection = new Selection(this, []);
 		this.selection = new Selection(this, []);
@@ -709,10 +132,10 @@ class RenderContext {
 				this.changeBrush(new DeleteBrush(this));
 			}
 			else if(event.key === "a") {
-				this.changeBrush(new NodeAddBrush(this));
+				this.changeBrush(new AddBrush(this));
 			}
 			else if(event.key === "s") {
-				this.changeBrush(new NodeSelectBrush(this));
+				this.changeBrush(new SelectBrush(this));
 			}
 			this.requestRedraw();
 		});
@@ -1233,11 +656,11 @@ class RenderContext {
 
 		// Debug help
 		infoLine("Change brush mode with (A)dd, (S)elect or (D)elete.");
-		if(this.brush instanceof NodeAddBrush) {
+		if(this.brush instanceof AddBrush) {
 			infoLine("Click to add terrain");
 			infoLine("Hold Q while scrolling to change brush type; hold W while scrolling to change brush size.");
 		}
-		else if(this.brush instanceof NodeSelectBrush) {
+		else if(this.brush instanceof SelectBrush) {
 			infoLine("Click to select, drag to move.");
 			infoLine("Hold Shift to select an entire object, hold Control to add to an existing selection.");
 		}
